refactor: separate main page and filter by tag
