+++ src/index.tsx
@@ -3,6 +3,7 @@ import App from './App';
 import { ActivePopupNoteContextProvider } from './context/ActivePopupNoteContext';
 import { EditNoteContextProvider } from './context/EditNoteContext';
 import './index.scss';
+import { FilterTagContextProvider } from './context/FilterTag';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -11,7 +12,9 @@ const root = ReactDOM.createRoot(
 root.render(
   <ActivePopupNoteContextProvider>
     <EditNoteContextProvider>
-      <App />
+      <FilterTagContextProvider>
+        <App />
+      </FilterTagContextProvider>
     </EditNoteContextProvider>
   </ActivePopupNoteContextProvider>
 );
